ActivityTaskPoller quits polling early
The ActivityTaskPoller fails to check for poll timeouts before logging information about the task.  As a result, it throws an exception, which is caught by the ActivityWorker, which just retries.  There's no real harm done, since the ActivityWorker retries the failed poll, but this cleans up the invalid log messages and lets the code behave as intended.