Explicitly require uuidtools
The uuidtools gem is used to generate UUIDs for workflow ids. The gem
was implicitly required via aws-sdk, but aws-sdk has dropped its
uuidtools dependency as of version 1.39.0 (by way of
aws/aws-sdk-ruby#461).

This gem may also wish to eliminate its dependency on uuidtools, but as
with aws-sdk, that would require either implementing a similar backport
of `SecureRandom.uuid` or bumping its aws-sdk requirement to at least
1.39.0 (or deliberately dropping support for older versions of ruby).